Update list of required files for the reading of Nortek wave data
Update list of required files for wave processing and print out informative message if needed about which files are missing. The addition of whr file is for a future requirement this it might be the only way to detect if a user has applied a compass offset during wave processing.