tools to get/send testfiles to imos-toolbox bucket
The data can be reach by several tools,
but in the case a user don't know the tools
a script in python is provided to get the files.

Only authorized POs can send data, so I'm keeping
a simple send script in shell for simplicity and to
allow easy parallelisation.