CPHL name/comment in AquatrackaIII & conflict msg
The comment is extended to identify
analogic conversion and explicitly
report scale/offset factors.

The commit also add a safety warning msg in
case of name conflicts.

This may be the case if the user combine
this routine with convertSBEcnvVar and there
is a CPHL already defined.

I didn't enumerate CPHL in this corner case
since this may never occur - the previous code
do not handle multiple CPHL variables anyway.