feat(Parser): generic parser machinery
This includes all of the generic parser machinery.
So far it includes support for non-structure headers - mostly
Staroddi instruments.

argstruct- a wrapper to create a particular struct.

detectEncoding - detect the encoding of a file.

readLinesWithRegex - read lines of a file based on a regex signature.

readHeaderDict- read header lines based on a function and arguments.

translateHeaderText - translate header text to data based on rules.

mapdataNames - Map header names to variable names.

HeaderParser - top-level function to parse headers of a files based on
an ensemble of rules.

DataParser - top-level function to parse data fields of a files based
on an ensemble of rules.