feat(Parsers): Staroddi Generic parser & tests
This include a new Staroddi parser with the new
Generic Parser machinery.

The machinery requires a set of rules and a parser
function. See both files for help in usage.

The parser allow processing of:

Starmon Mini v000
Starmon Mini v001
Starmon DST Tilt
Starmon DST CTD

See testStartoddi.m

Given the non-structure nature of the headers,
the set of rules is a bit big so StaroddiRules
is heavily commented. This is also the case
since this is the first example of the generic parser machinery.