feat(QC): include a generic numericInputButton
this is a class with a basic editable ui text button.

The class allows a validation function callback when
the user finishes to type, which restores the
editable field to a default value if a wrong input
is provided.

Useful to avoid a while/callback wrapper to ask the user
for valid input or unexpected aborts.