feat(qc): update imosTimeSeriesSpikeQC
This completely modifies the previous imosTimeSeriesSpikeQC
functionality.

1. Before we spawn one window per variable for spikeQC
selection. Now we spawn a single window that allows
multi variable/test selection.
This removes the requirement for globals previously
used, at the expense of leaking a particular condition in qcFilterMain.

2. The new UI window is handle by a class - spikeWindow.

3. The new class, spikeWindow, provides a preview window class to
play and interact with the select tests and their parameters.

The other functions are utilities that were updated included
for the above windows to work properly. The numericInputButton
was updated to allow a properly defined callbacks (LaterCallback).