fix(graph): fallback for missing the image toolbox
The problem:

The `drawrectangle` function, which
extract user interactive click/drag
positions, is a proprietary function
of the image processing toolbox.
Hence, usage is limited to users that
are licensed to it.

The fix:

Include a fallback function for users
without the toolbox.

How is fixed:

The select_points function is restored
on the repository, modified as
a closure to handle the positional
extraction. It will call
`drawrectangle` if available, otherwise
it will extract the positions with
the previous/old code.