feat(utility): tests as docstrings
This commit contains functions
that allow tests/examples
described in docstrings to be
evaluated. Thus, this allows
unit-tests to be written as docstrings.

The implementation is based
on the recent docstring template
used across the toolbox
where we have an `%Example:` block.

`test_docstring` is used to evaluate
the Example block of a single
function/class.

`check_docstrings` is used for
recursive evaluation of files
at a folder level.

`read_until_match` is used
for the docstring reading/block find,
while `comment_eval_wrapper`
is used as the closure for
the string evaluation.