feat(Util): getindex & wrappers for type validation
`getindex` is a wrapper to get the index
value of common matlab types.
It serves the purpose
to be used inside an anonymous function,
and to extract a fieldname of a struct
with the respective fieldname index.

`isindex` evaluate if the argument
is a valid array index:

1. non-zero positive integer; or
2. non-zero positive integer array; or
3. logical array with at least 1 true value;

`issize` is just a a checker for a
positive integer numeric array.

Finally, simple wrappers to check the types within
a cell are:

`iscellfh` - check items are function-handles.
`iscellnumeric` - ditto, but for any numeric value.
`iscellstruct` - ditto, but for structs.
`iscellsize` - ditto, but for positive, non-zero,
and row vector only integers.