feat(bugfix): test_docstring/check_docstrings
This commit fixes the too many open files
problem in `test_docstring` by adding a
missing `fclose` on the file id.

This can happen during long sessions or
repeated calls to `test_docstring`/`check_docstrings`.

A small fix is also provided in `check_docstrings`
where a call to an invalid/empty folder was not
generating an error.