feat(Parser): OceanContour netcdf & mat parser
This commit introduces the `oceanContourParser`,
supporting the netCDF and matfiles from the Nortek
OceanContour software.

The parser was based on the `signatureParser` code
and import the same metadata and variables.

The `OceanContour` class contains all methods
related to the parser, including variable
and attribute mappings between entities,
as well as the actual reader.

At the moment, support is available mostly for
ping average data ("Avg").  More test files
are needed for other types of output.