feat(Util): mem usage and double/cell comparisons
This commit includes 3 new utilities:

1. `isequal_tol` perform comparisons of double/single
arrays based on decimal precision.
The rationale is to estimate percentage of total
members that are equal based on floating point quantisation.
Percentages are required since quantisation/rounding
can lead to differences up to +-1 decimal range chosen.

2. `isinside` is just a wrapper that estimate if a certain
   object is inside another. This is currently implemented
   only for cells.

3. `memuse` is a array memory usage estimator in MiB.
Instead of creating a large array and asking the size.
we just calculate from the size and type.