fix(adcpBinMappingPP): allow downward adcp
This fix a bug for downlooking adcp
introduced when we changed the interpolation function
and cleanup a variable not used.

A constraint for the interpolation function is that
the values along all dimensions should be increasing.

This fix just shift the value within the function closure
so we can perform interpolation of downlooking bins.