feat(IMOS): schema evaluation for PP/QC
This implements a simple function based schema validation
for PP/QC routines. The idea is to externalise some long
verification calls and inspections of individual routines
and turn the actual code that perform the PP/QC cleaner.

The schemas must be named exactly as functions within the
`+schema` folder.

`IMOS.validate_dataset` perform a schema validation of a function.

For practical examples, see
imosEchoIntensitySetQC & imosSurfaceDetectionByDepthSetQC.