feat(Util): remove wrapTo360 dependency
This fixes another case of missing dependency for the stand alone
binary.

The function `wrapTo360` is not present in runtime v96, so
the standalone binary is broken when using this function.

Moreover, the use here is quite unnecessary, since
requires back-forth conversion for the 360 degree case.

The conversion is now direct, given the atan2d is bounded by
-180-180 degrees anyway.