fix(tests): fix docstrings
This fix both description and tests in the
SpikeClassifiers routines, since they
were outdated.

In `imosSpikeClassifierBurstHampel`, the description
for madfactor was wrong (wiki was correct) and
the test for the "ignore burst" functionality
was not very explicit.

Within the `BurstRunningStats,NonBurstSavGolOTSU,RunningStats`
functions, the docstrings were outdated with the function
behaviour, which was to return the index of spikes (a-la find).