fix(Parser): workhorseParse U and V swap bug
This fixes bug #741, where both the code
and docstring tests were assigning U (V)
velocity to beam number 2 (1) for
Teledyne workhorses ENU ADCP datasets.

Apart from the code/test fix, a new `xunit`
regression test is now provided,
where we compare the results of the reading with
a previous file that was read with
version 2.6.9 (before the
changes were introduced).

Finally, the commits also include
a small refactoring in `import_mappings`
so variable assignment is more explicit,
and in `workhorseParse` with better
error msg at the import step.