fix(Parser): wave magnetic variable renaming
This fixes #743 - the undefined variable bug
introduced when refactoring the `workhorseParse`.

The old variable name `magExt` was renamed to
`magdec_name_extension`, but the wave-related
code was unchanged, leaving the old variable
name in the source.

As well, and undefined variable
`magdec_name_comment` was being used
before being defined.

Now, all the `magExt` are renamed, and
we access the respective comment in
the wave-related code via
the `magdec_attrs` structure.