feat(IMOS): allow multiple attributes args
This expands the functionality of `IMOS.gen_dimensions`
and `IMOS.gen_variables`` to accept multiple attribute
arguments.

Previously all the arguments in the functions were
broadcasted to all dimensions/variables.

Now we also accept: a. varargin is as a flat cell
and b. cell of cells as varargin.

Option b. will assign the individual cell in order of the
dimension/variable assignment, skipping missing or invalid
entries.

For usage, see the updated docstrings.